them theme, vote, chinh lai db cho phu hop (file export)

+++ SocialNetworkBE/src/main/java/com/tlqt/repositories/PostRepository.java
@@ -20,5 +20,5 @@ public interface PostRepository{
     Post getPostById(int id);
     void deletePostById(int postId);
     List<Post> getPostsByAuthorId(int authorId);
-    int countPostsByUserId(int userId);
+    long countPostsByUserId(int userId);
 }
